requirements.txt
selenium==4.8.2
pytest~=7.2.1
pytest-soft-assertions~=0.1.2
pytest-xdist~=3.2.0
pytest-html~=3.2.0
allure-pytest~=2.12.0
behave~=1.2.6
allure-behave~=2.12.0
webdriver-manager~=3.8.5
pandas
pause
PyYAML~=5.4.1
requests~=2.28.2